+++ tests/server/validators/getValidConfig.ts
@@ -25,7 +25,8 @@ export const shouldGetValidConfig = () => {
         ["valid .json file", "./config.json", jsonFile],
         ["valid folder", "./", { ...jsFile, ...jsonFile }],
       ])('If %s path is passed', (_condition, configPath, expected) => {
-        const config = getValidConfig(configPath as string, path.join(__dirname, "../../mock/config"));
+        const rootPath = path.join(__dirname, "../../mock/config");
+        const config = getValidConfig(configPath as string, { rootPath });
         const expectedConfig = { ...Defaults.Config, ...expected };
         expect(config).toEqual(expectedConfig);
       });
@@ -41,7 +42,7 @@ export const shouldGetValidConfig = () => {
         ["base is empty string", { base: "" }, {}],
         ["base is '/' ", { base: "/" }, {}],
         ["base is '/api' ", { base: "/api" }, { base: "/api" }],
-        ["staticDir is not a valid path", { root: path.join(__dirname, "../../mock"), staticDir: "/mock" }, { root: path.join(__dirname, "../../mock"), staticDir:"C:\\mock" }],
+        ["staticDir is not a valid path", { root: path.join(__dirname, "../../mock"), staticDir: "/mock" }, { root: path.join(__dirname, "../../mock"), staticDir: "C:\\mock" }],
         ["staticDir is a valid path", { root: path.join(__dirname, "../../mock"), staticDir: "../../public" }, { root: path.join(__dirname, "../../mock"), staticDir: path.join(__dirname, "../../mock", "../../public") }],
         ["host is a empty string", { host: "" }, {}],
         ["host is not a string", { host: 129 }, {}],